refactor: rename new state module to state
